refactor: Extract PlannedTransactionsPage header in its own component
